requirements.txt
Django==3
psycopg2-binary==2.9.9
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-rest-auth==0.9.5
django-extensions==3.1.5 
